OpenSSL: Negotiate TLS version by default
If no $version was given to OpenSSL.new, it would use the highest
TLS version implemented in the OpenSSL library. This is a problem
when you access, via e.g. the WWW module which gives you no access
to the OpenSSL instance, a server which only implements older TLS.

Now, a version negotiating TLS method is used instead if no
explicit $version is requested. Prior to OpenSSL 1.1.0, this
was SSLv23_method() and was then renamed to TLS_method(). [^1]

The OpenSSL documentation SSL_CTX_new(3) recommends to use the
version-flexible methods in applications.

Also adapt the 01-basic.t test #3. With OpenSSL 1.1.0, there is
a special method version 0x10000 indicating version negotiation.
Prior to that, the version was set to the same as the highest
implemented method. [^2,3]

[^1] https://wiki.openssl.org/index.php/SSL/TLS_Client#Context_Setup
[^2] Added TLSv1.1: https://github.com/openssl/openssl/blob/637f374ad49d5f6d4f81d87d7cdd226428aa470c/ssl/ssl_locl.h#L673
[^3] Added TLSv1.2: https://github.com/openssl/openssl/blob/49f6cb968ff63793f6671d9026fb2a7034dad79a/ssl/ssl_locl.h#L694