Test installation and staging
When doing CI tests it is important to also test that the install
itself works. However, due to how precompilation works with staging
repositories the fastest way to both run the tests and test the
actual install is to run `zef install .`. Doing so avoids an extra
round of precompilation over doing a test step followed by an
install step.